pyproject.toml
[project]
name = "saml_registry"
version = "0.1.0"

dependencies = [
	"authzsvc-api-python-wrapper==1.3",
	"pyOpenSSL~=25.1.0",
	"defusedxml~=0.7.1",
	"pydantic-settings~=2.11.0",
	"Jinja2~=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"